rename mssqldb module as sqldb
