+++ pkg/services/sqldb/arm_template_new_server.go
@@ -1,4 +1,4 @@
-package mssqldb
+package sqldb
 
 // nolint: lll
 var armTemplateNewServerBytes = []byte(`
