+++ pkg/services/sqldb/config.go
@@ -1,4 +1,4 @@
-package mssqldb
+package sqldb
 
 import (
 	"encoding/json"
