+++ pkg/services/sqldb/deprovision.go
@@ -1,4 +1,4 @@
-package mssqldb
+package sqldb
 
 import (
 	"context"
