+++ pkg/services/sqldb/provision.go
@@ -1,4 +1,4 @@
-package mssqldb
+package sqldb
 
 import (
 	"context"
