+++ pkg/services/sqldb/unbind.go
@@ -1,4 +1,4 @@
-package mssqldb
+package sqldb
 
 import (
 	"fmt"
